Address the comments
