--- helix-core/src/test/java/org/apache/helix/integration/task/TestTaskErrorMaxRetriesQuotaRelease.java
@@ -1,63 +1,0 @@
-package org.apache.helix.integration.task;
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import org.apache.helix.TestHelper;
-import org.apache.helix.task.JobConfig;
-import org.apache.helix.task.JobQueue;
-import org.apache.helix.task.TaskState;
-import org.apache.helix.task.TaskUtil;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-import com.google.common.collect.ImmutableMap;
-
-public class TestTaskErrorMaxRetriesQuotaRelease extends TaskTestBase {
-
-  @BeforeClass
-  public void beforeClass() throws Exception {
-    _numNodes = 1;
-    _numPartitions = 100;
-    super.beforeClass();
-  }
-
-  @AfterClass
-  public void afterClass() throws Exception {
-    super.afterClass();
-  }
-
-  @Test
-  public void testTaskErrorMaxRetriesQuotaRelease() throws Exception {
-    String jobQueueName = TestHelper.getTestMethodName();
-    String jobName = "JOB0";
-    JobConfig.Builder jobBuilder = JobConfig.Builder.fromMap(WorkflowGenerator.DEFAULT_JOB_CONFIG)
-        .setMaxAttemptsPerTask(2).setWorkflow(jobQueueName).setFailureThreshold(100000)
-        .setJobCommandConfigMap(
-            ImmutableMap.of(MockTask.JOB_DELAY, "10", MockTask.FAILURE_COUNT_BEFORE_SUCCESS, "10"));
-
-    JobQueue.Builder jobQueue = TaskTestUtil.buildJobQueue(jobQueueName);
-    jobQueue.enqueueJob(jobName, jobBuilder);
-
-    _driver.start(jobQueue.build());
-    _driver.pollForJobState(jobQueueName, TaskUtil.getNamespacedJobName(jobQueueName, jobName),
-        TaskState.COMPLETED);
-  }
-}
